feat: allow use of env variables for db credentials
This commit allows getting the password, username and/or url for the
database from the respective OSOC_DB_PASSWORD, OSOC_DB_USERNAME and
OSOC_DB_URL env variables if they are set. In the case that they are not
set default values from the "application.properties" file will be used
just like before this commit.

Implementation details:
The "application.properties" file is loaded lazily into the
applicationProperties field when it is needed and reused if any other
env variables were not set.